Move errors to internal.compiletime
