+++ chimney/src/main/scala/io/scalaland/chimney/internal/compiletime/DerivationResult.scala
@@ -1,7 +1,5 @@
 package io.scalaland.chimney.internal.compiletime
 
-import io.scalaland.chimney.internal.{PatcherDerivationError, TransformerDerivationError}
-
 import scala.collection.compat.*
 import scala.util.control.NonFatal
