+++ chimney/src/main/scala/io/scalaland/chimney/internal/compiletime/derivation/transformer/ResultOps.scala
@@ -1,10 +1,10 @@
 package io.scalaland.chimney.internal.compiletime.derivation.transformer
 
-import io.scalaland.chimney.internal.compiletime.DerivationResult
-import io.scalaland.chimney.internal.{
+import io.scalaland.chimney.internal.compiletime.{
   AmbiguousCoproductInstance,
   CantFindCoproductInstanceTransformer,
   CantFindValueClassMember,
+  DerivationResult,
   IncompatibleSourceTuple,
   MissingAccessor,
   MissingJavaBeanSetterParam,
